[Lens] Split up dimension panel code (#80423)
* [Lens] Split up dimension panel code

* Fix test failures

* Style updates
