+++ x-pack/plugins/lens/public/indexpattern_datasource/__mocks__/state_helpers.ts
@@ -17,4 +17,5 @@ export const {
   sortByField,
   hasField,
   updateLayerIndexPattern,
+  mergeLayer,
 } = actual;
