+++ x-pack/plugins/lens/public/indexpattern_datasource/dimension_panel/index.ts
@@ -5,3 +5,5 @@
  */
 
 export * from './dimension_panel';
+export * from './droppable';
+export * from './operation_support';
